fix(logout): fix show message confirmation
